refactor: Move aqua package under pkg/plugin
