+++ pkg/plugin/aqua/client/auth.go
