+++ pkg/plugin/aqua/client/client_integration_test.go
