+++ pkg/plugin/aqua/client/doc.go
@@ -1,0 +1,2 @@
+// The client package provides an HTTP client for selected Aqua Enterprise REST API endpoints.
+package client
