+++ pkg/plugin/aqua/doc.go
@@ -1,0 +1,2 @@
+// The aqua package provides primitives for working with Aqua Enterprise scanner.
+package aqua
