+++ pkg/plugin/aqua/scanner/cli/types.go
