+++ pkg/plugin/aqua/scanner/testdata/aqua_csp_report/with_vulnerabilities.json
